+++ js/examples/apps/demo-document-relevancy-experiment/CHANGELOG.md
@@ -1,5 +1,20 @@
 # demo-document-relevancy-experiment
 
+## 1.0.11
+
+### Patch Changes
+
+- @arizeai/phoenix-client@5.8.1
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [0738cb9]
+- Updated dependencies [d6d1953]
+  - @arizeai/phoenix-evals@0.9.0
+  - @arizeai/phoenix-client@5.8.0
+
 ## 1.0.9
 
 ### Patch Changes